+++ ContentPatcher/release-notes.md
@@ -18,6 +18,7 @@ When releasing a format change, don't forget to update the smapi.io/json schema!
 * The `FarmhouseUpgrade` token can now check either the current player (default) or the host player.
 * The `Enabled` field no longer allows tokens (in format version 1.21.0+).
 * Improved default `LogName` for patches with multiple `Target` or `FromFile` values.
+* Improved split-screen support.
 * Fixed changes through Generic Mod Config Menu not correctly updating dynamic tokens and `Enabled` fields.
 * Fixed `patch reload` command not reapplying format migrations to reloaded patches.
 * Fixed error patching `Data\Concessions` using `EditData`.